+++ R/sanitize.R
@@ -11,12 +11,8 @@ sanitize_plan <- function(plan, allow_duplicated_targets = FALSE){
       }
     }
   }
-  if ("trigger" %in% colnames(plan)){
-    plan$trigger <- parse_triggers(plan$trigger)
-  }
   plan$target <- repair_target_names(plan$target)
   plan <- plan[nzchar(plan$target), ]
-  plan$command[is.na(plan$command)] <- ""
   first <- c("target", "command")
   cols <- c(first, setdiff(colnames(plan), first))
   if (!allow_duplicated_targets) {
@@ -90,10 +86,4 @@ repair_target_names <- function(x){
   x <- gsub("^_", "", x)
   x[!nzchar(x)] <- "X"
   make.unique(x, sep = "_")
-}
-
-parse_triggers <- function(x){
-  x[is.na(x) | !nzchar(x)] <- "any"
-  assert_legal_triggers(x)
-  x
 }