+++ tests/testthat/test-triggers.R
@@ -2,145 +2,310 @@ drake_context("triggers")
 
 test_with_dir("empty triggers return logical", {
   skip_on_cran() # CRAN gets whitelist tests only (check time limits).
-  expect_identical(depends_trigger("x", list(), list()), FALSE)
+  expect_identical(depend_trigger("x", list(), list()), FALSE)
   expect_identical(command_trigger("x", list(), list()), FALSE)
   expect_identical(file_trigger("x", list(), list()), FALSE)
+  expect_identical(change_trigger("x", list(), list()), FALSE)
 })
 
-test_with_dir("triggers work as expected", {
+test_with_dir("triggers can be expressions", {
   skip_on_cran() # CRAN gets whitelist tests only (check time limits).
-  con <- dbug()
-  con$plan$trigger <- "missing"
-  con <- testrun(config = con)
-  expect_equal(outdated(config = con), character(0))
-
-  # Illegal trigger
-  con$plan$trigger[1] <- 5
-  expect_error(testrun(con))
-  con$plan$trigger[1] <- "missing"
-
-  # Change a command.
-  cmd <- con$plan$command[con$plan$target == "combined"]
-  con$plan$command[con$plan$target == "combined"] <-
-    "nextone + yourinput + 1"
-  for (trigger in setdiff(triggers(), c("always", triggers_with_command()))){
-    con$plan$trigger <- trigger
-    expect_equal(outdated(config = con), character(0))
-  }
-  for (trigger in triggers_with_command()){
-    con$plan$trigger[con$plan$target == "combined"] <- trigger
-    expect_equal(sort(outdated(config = con)),
-      sort(c("combined", "final", "drake_target_1")))
-  }
-  con$plan$command[con$plan$target == "combined"] <- cmd
-
-  # Destroy a file target.
-  file.rename("intermediatefile.rds", "tmp")
-  check_file <- function(con){
-    for (trigger in setdiff(triggers(), c("always", triggers_with_file()))){
-      con$plan$trigger <- trigger
-      expect_equal(outdated(config = con), character(0))
-    }
-    for (trigger in triggers_with_file()){
-      con$plan$trigger[con$plan$target == "drake_target_1"] <- trigger
-      expect_equal(sort(outdated(config = con)),
-        sort(c("final", "drake_target_1")))
-    }
-  }
-  check_file(con)
-
-  # Restore the file target.
-  file.rename("tmp", "intermediatefile.rds")
-  con$plan$trigger <- "always"
-  expect_equal(sort(outdated(config = con)), sort(con$plan$target))
-  for (trigger in setdiff(triggers(), "always")){
-    con$plan$trigger <- trigger
-    expect_equal(outdated(config = con), character(0))
-  }
-
-  # Corrupt a file target
-  value <- readRDS("intermediatefile.rds") + 1
-  saveRDS(value, "intermediatefile.rds")
-  check_file(con)
-
-  # Nothing is built (global missing trigger)
-  con$plan$command[con$plan$target == "yourinput"] <- "1+2"
-  file.rename("intermediatefile.rds", "tmp")
-  con$plan$trigger <- NULL
-  con <- make(
-    con$plan, trigger = "missing",
-    envir = con$envir, verbose = TRUE, session_info = FALSE)
-  expect_equal(justbuilt(con), character(0))
-
-  # Global trigger is overridden
-  con$plan$trigger <- "missing"
-  con <- make(
-    con$plan, trigger = "command",
-    envir = con$envir, verbose = TRUE, session_info = FALSE)
-  expect_equal(justbuilt(con), character(0))
-
-  # 'always' trigger rebuilts up-to-date targets
-  con$plan$trigger <- "any"
-  con <- make(con$plan, envir = con$envir, session_info = FALSE)
-  out <- outdated(con)
-  expect_equal(out, character(0))
-  con$plan$trigger[con$plan$target == "final"] <- "always"
-  con2 <- make(
-    con$plan, parallelism = con$parallelism,
-    envir = con$envir, jobs = con$jobs, verbose = FALSE,
-    session_info = FALSE)
-  expect_equal(justbuilt(con2), "final")
-})
-
-test_with_dir("all triggers bring targets up to date", {
-  skip_on_cran() # CRAN gets whitelist tests only (check time limits).
-  for (trigger in triggers()){
-    clean(destroy = TRUE)
-    con <- dbug()
-    con$plan$trigger <- trigger
-    con <- make(
-      con$plan, parallelism = con$parallelism,
-      envir = con$envir, jobs = con$jobs, verbose = FALSE,
-      session_info = FALSE)
-    expect_equal(sort(justbuilt(con)), sort(con$plan$target))
-    con$plan$trigger <- NULL
-    out <- outdated(con)
-    expect_equal(out, character(0))
+  plan <- drake_plan(x = 1)
+  plan$trigger <- expression(trigger(condition = TRUE))
+  for (i in 1:3) {
+    config <- make(
+      plan,
+      session_info = FALSE,
+      cache = storr::storr_environment()
+    )
+    expect_equal(justbuilt(config), "x")
   }
 })
 
-# Similar enough to the triggers to include here:
-test_with_dir("make(..., skip_imports = TRUE) works", {
+test_with_dir("triggers in plan override make(trigger = whatever)", {
   skip_on_cran() # CRAN gets whitelist tests only (check time limits).
-  con <- dbug()
-  verbose <- max(con$jobs) < 2 &&
-    targets_setting(con$parallelism) == "parLapply"
-  suppressMessages(
-    con <- make(
-      con$plan, parallelism = con$parallelism,
-      envir = con$envir, jobs = con$jobs, verbose = verbose,
-      hook = silencer_hook,
-      skip_imports = TRUE,
-      session_info = FALSE
+  saveRDS(1, "file.rds")
+  plan <- drake_plan(
+    x = readRDS(file_in("file.rds")),
+    y = target(
+      readRDS(file_in("file.rds")),
+      trigger(file = TRUE)
+    ),
+    strings_in_dots = "literals"
+  )
+  config <- make(plan)
+  expect_equal(sort(justbuilt(config)), c("x", "y"))
+  saveRDS(2, "file.rds")
+  expect_equal(sort(outdated(config)), c("x", "y"))
+  config <- make(plan, trigger = trigger(file = FALSE))
+  expect_equal(justbuilt(config), "y")
+})
+
+test_with_dir("change trigger on a fresh build", {
+  skip_on_cran() # CRAN gets whitelist tests only (check time limits).
+  saveRDS(1, "file.rds")
+  plan <- drake_plan(
+    x = target(1 + 1, trigger(
+      condition = FALSE,
+      command = FALSE,
+      depend = FALSE,
+      file = FALSE,
+      change = readRDS("file.rds"))
+    ),
+    strings_in_dots = "literals"
+  )
+  config <- make(plan, session_info = FALSE)
+  expect_equal(justbuilt(config), "x")
+  config <- make(plan, session_info = FALSE)
+  expect_equal(justbuilt(config), character(0))
+  saveRDS(2, "file.rds")
+  config <- make(plan, session_info = FALSE)
+  expect_equal(justbuilt(config), "x")
+})
+
+test_with_dir("trigger() function works", {
+  skip_on_cran() # CRAN gets whitelist tests only (check time limits).
+  x <- 1
+  y <- trigger(
+    condition = 1 + 1,
+    command = TRUE,
+    depend = FALSE,
+    file = FALSE,
+    change = sqrt(!!x)
+  )
+  z <- list(
+    condition = quote(1 + 1),
+    command = TRUE,
+    depend = FALSE,
+    file = FALSE,
+    change = quote(sqrt(1))
+  )
+  expect_equal(y, z)
+})
+
+test_with_dir("can detect trigger deps", {
+  skip_on_cran() # CRAN gets whitelist tests only (check time limits).
+  f <- function(x){
+    identity(x)
+  }
+  plan <- drake_plan(
+    x = target(
+      command = 1 + 1,
+      trigger = trigger(
+        condition = f(FALSE),
+        command = FALSE,
+        file = FALSE,
+        depend = TRUE,
+        change = NULL
+      )
+    ),
+    strings_in_dots = "literals"
+  )
+  config <- drake_config(
+    plan, session_info = FALSE, cache = storr::storr_environment(),
+    log_progress = TRUE)
+  expect_equal(dependencies("x", config), "f")
+  expect_equal(outdated(config), "x")
+  make(config = config)
+  expect_equal(justbuilt(config), "x")
+  expect_equal(outdated(config), character(0))
+  make(config = config)
+  nobuild(config)
+  f <- function(x){
+    identity(x) || FALSE
+  }
+  expect_equal(outdated(config), "x")
+  make(config = config)
+  expect_equal(justbuilt(config), "x")
+})
+
+test_with_dir("triggers can be NA in the plan", {
+  skip_on_cran()
+  expect_silent(
+    config <- make(
+      drake_plan(x = target(1, NA)),
+      session_info = FALSE,
+      cache = storr::storr_environment(),
+      verbose = FALSE
     )
   )
+  expect_equal(justbuilt(config), "x")
+})
+
+test_with_dir("trigger components react appropriately", {
+  skip_on_cran()
+  scenario <- get_testing_scenario()
+  e <- eval(parse(text = scenario$envir))
+  jobs <- scenario$jobs
+  parallelism <- scenario$parallelism
+  caching <- scenario$caching
+  eval(
+    quote(f <- function(x){
+      1 + x
+    }),
+    envir = e
+  )
+  writeLines("1234", "report.Rmd")
+  saveRDS(1, "file.rds")
+  saveRDS(1, "change.rds")
+  saveRDS(TRUE, "condition.rds")
+  plan <- drake_plan(
+    missing = target(
+      "",
+      trigger(command = FALSE, depend = FALSE, file = FALSE)
+    ),
+    condition = target(
+      "",
+      trigger(
+        condition = readRDS("condition.rds"),
+        command = FALSE, depend = FALSE, file = FALSE
+      )
+    ),
+    command = target(
+      "",
+      trigger(command = TRUE, depend = FALSE, file = FALSE)
+    ),
+    depend = target(
+      "",
+      trigger(command = FALSE, depend = TRUE, file = FALSE)
+    ),
+    file = target(
+      "",
+      trigger(command = FALSE, depend = FALSE, file = TRUE)
+    ),
+    change = target(
+      "",
+      trigger(
+        change = readRDS("change.rds"),
+        command = FALSE, depend = FALSE, file = FALSE
+      )
+    ),
+    strings_in_dots = "literals"
+  )
+  plan$command <- paste0("
+    knitr_in(\"report.Rmd\")
+    out <- f(readRDS(file_in(\"file.rds\")))
+    saveRDS(out, file_out(\"out_", plan$target, ".rds\"))
+    out
+  ")
+  config <- make(
+    plan, envir = e, jobs = jobs, parallelism = parallelism,
+    verbose = FALSE, caching = caching, session_info = FALSE
+  )
+  expect_equal(sort(justbuilt(config)), sort(config$plan$target))
+  simple_plan <- plan
+  simple_plan$trigger <- NULL
+  simple_config <- make(
+    simple_plan, envir = e, jobs = jobs, parallelism = parallelism,
+    verbose = FALSE, caching = caching, session_info = FALSE
+  )
+
+  # Condition trigger
+  for (i in 1:2){
+    expect_equal(sort(outdated(config)), "condition")
+    make(config = config)
+    expect_equal(sort(justbuilt(config)), "condition")
+  }
+  saveRDS(FALSE, "condition.rds")
+  expect_equal(outdated(simple_config), character(0))
+  expect_equal(outdated(config), character(0))
+  for (i in 1:2){
+    make(config = config)
+    nobuild(config)
+  }
+
+  # Change trigger
+  saveRDS(2, "change.rds")
+  expect_equal(sort(outdated(config)), "change")
+  expect_equal(outdated(simple_config), character(0))
+  make(config = config)
+  expect_equal(sort(justbuilt(config)), "change")
+  expect_equal(outdated(config), character(0))
+  expect_equal(outdated(simple_config), character(0))
+
+  # File trigger: input files
+  saveRDS(2, "file.rds")
+  expect_equal(sort(outdated(config)), "file")
+  make(config = config)
+  expect_equal(sort(justbuilt(config)), "file")
   expect_equal(
-    sort(cached()),
-    sort(c("\"intermediatefile.rds\"", con$plan$target))
+    sort(outdated(simple_config)),
+    sort(setdiff(config$plan$target, "file"))
   )
+  expect_equal(outdated(config), character(0))
 
-  # If the imports are already cached, the targets built with
-  # skip_imports = TRUE should be up to date.
-  make(con$plan, verbose = FALSE, envir = con$envir, session_info = FALSE)
-  clean(list = con$plan$target, verbose = FALSE)
-  suppressMessages(
-    con <- make(
-      con$plan, parallelism = con$parallelism,
-      envir = con$envir, jobs = con$jobs, verbose = verbose,
-      hook = silencer_hook,
-      skip_imports = TRUE, session_info = FALSE
-    )
+  # File trigger: knitr files
+  writeLines("5678", "report.Rmd")
+  expect_equal(sort(outdated(config)), "file")
+  make(config = config)
+  expect_equal(sort(justbuilt(config)), "file")
+  expect_equal(
+    sort(outdated(simple_config)),
+    sort(setdiff(config$plan$target, "file"))
   )
-  out <- outdated(con)
-  expect_equal(out, character(0))
+  expect_equal(outdated(config), character(0))
+
+  # File trigger: output files
+  for (target in plan$target){
+    saveRDS("1234", paste0("out_", target, ".rds"))
+  }
+  expect_equal(sort(outdated(config)), "file")
+  make(config = config)
+  expect_equal(sort(justbuilt(config)), "file")
+  expect_equal(
+    sort(outdated(simple_config)),
+    sort(setdiff(config$plan$target, "file"))
+  )
+  expect_equal(outdated(config), character(0))
+
+  # Done with the change trigger
+  plan <- plan[1:5, ]
+  config <- drake_config(
+    plan, envir = e, jobs = jobs, parallelism = parallelism,
+    verbose = FALSE, caching = caching, log_progress = TRUE,
+    session_info = FALSE
+  )
+  simple_plan <- simple_plan[1:5, ]
+  simple_config <- drake_config(
+    simple_plan, envir = e, jobs = jobs, parallelism = parallelism,
+    verbose = FALSE, caching = caching, log_progress = TRUE,
+    session_info = FALSE
+  )
+  make(config = simple_config)
+
+  # Command trigger
+  config$plan$command <- simple_config$plan$command <- paste0("
+    knitr_in(\"report.Rmd\")
+    out <- f(1 + readRDS(file_in(\"file.rds\")))
+    saveRDS(out, file_out(\"out_", plan$target, ".rds\"))
+    out
+  ")
+  expect_equal(sort(outdated(config)), "command")
+  make(config = config)
+  expect_equal(sort(justbuilt(config)), "command")
+  expect_equal(
+    sort(outdated(simple_config)),
+    sort(setdiff(config$plan$target, "command"))
+  )
+  expect_equal(outdated(config), character(0))
+  make(config = simple_config)
+  expect_equal(outdated(config), character(0))
+
+  # Depend trigger
+  eval(
+    quote(f <- function(x){
+      2 + x
+    }),
+    envir = e
+  )
+  expect_equal(sort(outdated(config)), "depend")
+  make(config = config)
+  expect_equal(sort(justbuilt(config)), "depend")
+  expect_equal(
+    sort(outdated(simple_config)),
+    sort(setdiff(config$plan$target, "depend"))
+  )
+  expect_equal(outdated(config), character(0))
+  make(config = simple_config)
+  expect_equal(outdated(config), character(0))
 })